# Optimize `Object#instance_values`